Team 27 week 7 project checkpoint
I have implemented and trained DDIMs from scratch on numerous datasets,
wrote a relatively accessible introduction to DDPMs vs DDIMs and why the
latter is great for inversion, and implemented an optimization-based
method for inversion.

See the current project write-up for more details. The GitHub repository
to my code is also included there.